substantial refactor
adhereing to the framework proposed in "funcional core, imperative
shell", I tried to break out everything that's not stateless into its
own functions. Eventually these may become packages or files (I used to
think I understood the difference but now I'm not so sure). Error
handling is wonky, and is definitely a TODO for the future.