Treat 400/500 HTTP responses from Zulip as errors
Per [issue 55], Zulip will occasionally refuse a message. We weren't
passing this further up the stack, or even logging the status code.

Log the status code along with the body, and return an error up the
stack, if the Zulip request gets a user error (4xx) or server error
(5xx).

[issue 55]: https://github.com/recursecenter/pairing-bot/issues/55