Perform conversion from Firestore doc to Recurser
I've been observing some odd happening in how ID is handled w/rt its
duplication, as the key of the "recursers" collection, and as a value
in each collection.

Change the "deserialize" routine to be
`func (*DocumentSnapshot) (Recurser, Error)`,
which allows the user ID to be derived from the key (== the document
ID).