refactor: Extract assert package from existing tests
After the third copy-paste of assertEqual, I think it's time to make an
internal test assertion library.

This is similar to [testify]'s assert package, mostly because it's the
one I'm most familiar with.

This now uses [go-cmp] instead of `reflect.DeepEqual` for comparing
values. We get free diff reports out of it too!

I added the package to a new `internal` directory to make it clear that
this isn't part of the application itself. It also prevents it from
being imported from anywhere else, even though there's no reason to
import from this module anyway.

[go-cmp]: https://github.com/google/go-cmp
[testify]: https://github.com/stretchr/testify