refactor: Extract AppEngine Cron source check into middleware
This de-duplicates the code we have to check the AppEngine Cron request
header and pushes it into the HTTP router and out of the app logic.

As a convenient side effect, this also ensures that any cron job failure
produces an HTTP 500 response (indicating failure to AppEngine) and an
ERROR-severity log line (which we're starting to use for alerts).